yolo_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.5.0
torchvision==0.20.0
ultralytics>=8.3.0
python-multipart==0.0.6
pillow>=11.1.0
opencv-python-headless==4.10.0.84
numpy>=1.26.0
pydantic==2.5.3
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
